add image handler class
